Updated controllers to utilize the user id found in the req.user after authentication. Added a getCurrentUser function to save users from having to log in after reloading the page

+++ server/controllers/applicationController.js
@@ -3,7 +3,9 @@ const Application = require('../models/Application') // Application schema from 
 // Get all applications
 exports.getAllApplications = async (req, res) => {
     try {
-        const applications = await Application.find().sort({ createdAt: -1 });
+        const applications = await Application
+            .find({ user: req.user.id })
+            .sort({ createdAt: -1 });
         res.json(applications);
     } catch (error) {
         res.status(500).json({ message: 'Error fetching applications', error: error.message });
@@ -16,6 +18,7 @@ exports.createApplication = async (req, res) => {
         const { company, position, status, notes, followUpDate } = req.body;
 
         const newApplication = new Application({
+            user: req.user.id,
             company,
             position,
             status,
@@ -33,6 +36,15 @@ exports.createApplication = async (req, res) => {
 // Update application
 exports.updateApplication = async (req, res) => {
     try {
+        const application = await Application.findOne({
+            _id: req.params.id,
+            user: req.user.id
+        });
+        
+        if(!application){
+            return res.status(404).json({ message: "Application not found." });
+        }
+        
         const updatedApplication = await Application.findByIdAndUpdate(
             req.params.id,
             req.body,
@@ -52,6 +64,15 @@ exports.updateApplication = async (req, res) => {
 // Delete application
 exports.deleteApplication = async (req, res) => {
     try {
+        const application = await Application.findOne({
+            _id: req.params.id,
+            user: req.user.id
+        });
+        
+        if(!application){
+            return res.status(404).json({ message: "Application not found." });
+        }
+
         const deletedApplication = await Application.findByIdAndDelete(req.params.id);
 
         if(!deletedApplication) {
@@ -64,3 +85,17 @@ exports.deleteApplication = async (req, res) => {
         res.status(500).json({ message: 'Error deleting application', error: error.message });
     }
 };
+
+exports.getCurrentUser = async(req, res, next) => {
+    try {
+        const user = await User.findById(req.user.id).select('-password');
+
+        if(!user) {
+            return  res.status(404).json({ message: 'User not found. '});
+        }
+
+        res.json({ user });
+    } catch (error) {
+        next(error);
+    }
+}
